requirements.txt
streamlit
google-generativeai
graphviz
Pillow
